- make some variations to registration_deadline (should not affect us by the time we demo this) - added trigger to update start and end date of course offering - fix view_manager_report, promote_courses and pay_salary

+++ deliverables/schema.sql
@@ -30,7 +30,7 @@ CREATE TABLE Employees (
 CREATE TABLE Pay_slips (
     eid INTEGER,
     payment_date DATE,
-    amount INTEGER,
+    amount NUMERIC(12,2),
     num_work_hours INTEGER,
     num_work_days INTEGER,
     PRIMARY KEY(eid, payment_date), -- Weak entity set
